refactor: decouple Lark notification logic from sales order service into a dedicated notifier module
