+++ src/services/erp/selling/sales-order/sales-order.js
@@ -1,24 +1,22 @@
 import * as Sentry from "@sentry/cloudflare";
 import FrappeClient from "src/frappe/frappe-client";
 import { convertIsoToDatetime } from "src/frappe/utils/datetime";
-import LarksuiteService from "services/larksuite/lark";
 import Database from "src/services/database";
 import AddressService from "src/services/erp/contacts/address/address";
 import ContactService from "src/services/erp/contacts/contact/contact";
 import CustomerService from "src/services/erp/selling/customer/customer";
-import { composeOrderUpdateMessage, composeSalesOrderNotification, findMainOrder, isMissingJewelrySerial, extractPromotions } from "services/erp/selling/sales-order/utils/sales-order-notification";
-import { validateSalesOrder } from "services/erp/selling/sales-order/utils/sales-order-validator";
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc.js";
-import { CHAT_GROUPS } from "services/larksuite/group-chat/group-management/constant";
-import { fetchSalesOrdersFromERP, saveSalesOrdersToDatabase, ensureSelfReference, getLeadSource, fetchAndNormalizeAttachments, calculateGroupPayments } from "src/services/erp/selling/sales-order/utils/sales-order-helpers";
+import {
+  fetchSalesOrdersFromERP,
+  saveSalesOrdersToDatabase,
+  ensureSelfReference,
+  getLeadSource,
+  syncHaravanFinancialStatus
+} from "src/services/erp/selling/sales-order/utils/sales-order-helpers";
 import { getRefOrderChain } from "services/ecommerce/order-tracking/queries/get-initial-order";
-import Larksuite from "services/larksuite";
-import { getOrderFinancials } from "services/haravan/orders/order-service/helpers/order-financials";
-import { ERPR2StorageService } from "services/r2-object/erp/erp-r2-storage-service";
-import HaravanAPI from "services/clients/haravan-client";
-import { retryQuery } from "src/services/utils/retry-utils";
 import { isTestOrder } from "services/utils/order-intercepter";
+import { sendSalesOrderNotificationToLark } from "services/erp/selling/sales-order/utils/sales-order-lark-notifier";
 
 dayjs.extend(utc);
 
@@ -137,7 +135,7 @@ export default class SalesOrderService {
       total_amount: haravanOrderData.total_price,
       grand_total: haravanOrderData.total_price,
       real_order_date: await this.getRealOrderDate(haravanOrderData.id) || dayjs(haravanOrderData.created_at).utc().add(7, "hour").format("YYYY-MM-DD"),
-      ref_sales_orders: await this.mapRefSalesOrder(haravanOrderData.id),
+      ref_sales_orders: await this.fetchAndMapSalesOrderReferences(haravanOrderData.id),
       tracking_number: trackingNumber
     };
     const order = await this.frappeClient.upsert(mappedOrderData, "haravan_order_id", ["items"]);
@@ -176,7 +174,7 @@ export default class SalesOrderService {
     };
   };
 
-  mapRefSalesOrder = async (refOrderId) => {
+  fetchAndMapSalesOrderReferences = async (refOrderId) => {
     try {
       const refOrders = await getRefOrderChain(this.db, Number(refOrderId), true);
 
@@ -235,313 +233,8 @@ export default class SalesOrderService {
     };
   };
 
-  async sendNotificationToLark(initialSalesOrderData, isUpdateMessage = false) {
-    let salesOrderData = structuredClone(initialSalesOrderData);
-    const dbConnection = { timeout: 30000, maxWait: 10000 };
-    const larkClient = await LarksuiteService.createClientV2(this.env);
-
-    salesOrderData.attachments = await fetchAndNormalizeAttachments(
-      this.frappeClient,
-      salesOrderData.name,
-      this.env.JEMMIA_ERP_BASE_URL
-    );
-
-    const haravanRefOrderId = salesOrderData.haravan_ref_order_id;
-
-    const { allRelatedOrders } = await this.getAllRelatedSalesOrders(salesOrderData.name, salesOrderData);
-
-    const splitOrderGroupId = salesOrderData.split_order_group;
-    const isSplitOrder = salesOrderData.is_split_order;
-
-    let childOrders = [];
-    if (splitOrderGroupId && Number(splitOrderGroupId) > 0 && isSplitOrder) {
-      const splitOrders = allRelatedOrders.filter(o =>
-        o.split_order_group === splitOrderGroupId &&
-        o.name !== salesOrderData.name &&
-        o.cancelled_status === "Uncancelled"
-      );
-
-      if (splitOrders && splitOrders.length > 0) {
-        // For each order, find its attachments
-        for (const splitOrder of splitOrders) {
-          const childOrder = await this.frappeClient.getDoc("Sales Order", splitOrder.name);
-          childOrder.attachments = await fetchAndNormalizeAttachments(
-            this.frappeClient,
-            childOrder.name,
-            this.env.JEMMIA_ERP_BASE_URL
-          );
-          childOrders.push(childOrder);
-        }
-      }
-    }
-
-    const { mainOrder, subOrders } = findMainOrder([salesOrderData, ...childOrders]);
-
-    salesOrderData = mainOrder;
-    childOrders = subOrders;
-
-    salesOrderData.items = salesOrderData.items.map((item) => ({
-      ...item,
-      parent_order_number: salesOrderData.order_number,
-      parent_grand_total: salesOrderData.grand_total
-    }));
-
-    // Compose all sales order into one
-    for (const childOrder of childOrders) {
-      salesOrderData.items.push(
-        ...childOrder.items.map((item) => ({
-          ...item,
-          parent_order_number: childOrder.order_number,
-          parent_grand_total: childOrder.grand_total
-        }))
-      );
-      salesOrderData.attachments.push(...childOrder.attachments);
-      salesOrderData.promotions.push(...childOrder.promotions);
-      salesOrderData.product_categories.push(...childOrder.product_categories);
-      salesOrderData.policies.push(...childOrder.policies);
-
-      salesOrderData.grand_total += childOrder.grand_total;
-      salesOrderData.discount_amount += childOrder.discount_amount;
-    }
-
-    const payments = calculateGroupPayments(salesOrderData, childOrders);
-    salesOrderData.paid_amount = payments.paid_amount;
-    salesOrderData.deposit_amount = payments.deposit_amount;
-
-    const customer = await this.frappeClient.getDoc("Customer", salesOrderData.customer);
-
-    const { isValid, message } = validateSalesOrder(salesOrderData, customer);
-    if (!isValid) {
-      return { success: false, message: message };
-    }
-
-    const allOrdersToProcess = [salesOrderData, ...childOrders];
-    const allPromotionNames = new Set();
-    allOrdersToProcess.forEach(order => {
-      extractPromotions(order).forEach(p => allPromotionNames.add(p));
-    });
-
-    let promotionData = [];
-    if (allPromotionNames.size > 0) {
-      promotionData = await this.frappeClient.getList("Promotion", {
-        filters: [["name", "in", Array.from(allPromotionNames)]],
-        fields: ["*"]
-      });
-    }
-
-    if (haravanRefOrderId && Number(haravanRefOrderId) > 0) {
-      // find the very first order in history
-      const refOrders = await getRefOrderChain(this.db, Number(salesOrderData.haravan_order_id));
-
-      if (!refOrders || refOrders.length === 0) {
-        return {
-          success: false,
-          message: `Không tìm thấy đơn gốc của đơn ${salesOrderData.order_number}`
-        };
-      }
-
-      const refOrderstNotificationOrderTracking = await this.db.erpnextSalesOrderNotificationTracking.findMany({
-        where: {
-          haravan_order_id: {
-            in: refOrders?.map(order => String(order.id))
-          }
-        },
-        orderBy: {
-          database_created_at: "asc"
-        }
-      });
-
-      if (refOrderstNotificationOrderTracking && refOrderstNotificationOrderTracking.length > 0) {
-        const currentOrderTracking = await this.db.erpnextSalesOrderNotificationTracking.findFirst({
-          where: {
-            order_name: salesOrderData.name
-          }
-        });
-
-        const isOrderTracked = !!currentOrderTracking;
-
-        let content = null;
-        let diffAttachments = null;
-        if (isOrderTracked) {
-          ({ content, diffAttachments } = await this.composeUpdateOrderContent(
-            currentOrderTracking.order_data,
-            salesOrderData,
-            promotionData
-          ));
-        } else {
-          content = await this.composeNewOrderContent(salesOrderData, customer, promotionData, true);
-        }
-
-        if (!content && !diffAttachments) {
-          return { success: true, message: "Không có gì thay đổi!" };
-        }
-
-        const isSendImagesSuccess = await this._sendAttachmentsToLark(
-          larkClient,
-          diffAttachments?.added_file,
-          refOrderstNotificationOrderTracking[0].lark_message_id,
-          CHAT_GROUPS.CUSTOMER_INFO.chat_id
-        );
-
-        const replyResponse = content && await larkClient.im.message.reply({
-          path: {
-            message_id: refOrderstNotificationOrderTracking[0].lark_message_id
-          },
-          data: {
-            receive_id: CHAT_GROUPS.CUSTOMER_INFO.chat_id,
-            msg_type: "text",
-            reply_in_thread: true,
-            content: JSON.stringify({
-              text: content
-            })
-          }
-        });
-
-        if ((content && replyResponse.msg === "success") || (isSendImagesSuccess.every(Boolean))) {
-          if (isOrderTracked) {
-            await retryQuery(() => this.db.$transaction(async (tx) => {
-              return tx.erpnextSalesOrderNotificationTracking.updateMany({
-                where: {
-                  uuid: currentOrderTracking.uuid
-                },
-                data: {
-                  order_data: {
-                    items: salesOrderData.items,
-                    attachments: salesOrderData.attachments,
-                    paid_amount: salesOrderData.paid_amount,
-                    deposit_amount: salesOrderData.deposit_amount
-                  }
-                }
-              });
-            }, dbConnection));
-            return { success: true, message: "Cập nhật đơn thành công!" };
-          }
-          await retryQuery(() => this.db.erpnextSalesOrderNotificationTracking.create({
-            data: {
-              lark_message_id: replyResponse.data.message_id,
-              order_name: salesOrderData.name,
-              haravan_order_id: salesOrderData.haravan_order_id,
-              order_data: {
-                items: salesOrderData.items,
-                attachments: salesOrderData.attachments,
-                paid_amount: salesOrderData.paid_amount,
-                deposit_amount: salesOrderData.deposit_amount
-              }
-            }
-          }));
-          return { success: true, message: "Thông báo đơn đặt lại thành công!" };
-        }
-
-        if (isOrderTracked) {
-          return { success: false, message: "Cập nhật đơn thất bại!" };
-        }
-        return { success: false, message: "Thông báo đơn đặt lại thất bại!" };
-      }
-    }
-
-    const notificationTracking = await this.db.erpnextSalesOrderNotificationTracking.findFirst({
-      where: {
-        order_name: salesOrderData.name
-      }
-    });
-
-    if (notificationTracking) {
-      const { content, diffAttachments } = await this.composeUpdateOrderContent(notificationTracking.order_data || {}, salesOrderData, promotionData);
-
-      if (!content && !diffAttachments) {
-        return { success: false, message: "Đơn hàng này đã được gửi thông báo từ trước đó!" };
-      }
-
-      const isSendImagesSuccess = await this._sendAttachmentsToLark(
-        larkClient,
-        diffAttachments?.added_file,
-        notificationTracking.lark_message_id,
-        CHAT_GROUPS.CUSTOMER_INFO.chat_id
-      );
-
-      // Reply to the root message in the group chat
-      const replyResponse = content && await larkClient.im.message.reply({
-        path: {
-          message_id: notificationTracking.lark_message_id
-        },
-        data: {
-          receive_id: CHAT_GROUPS.CUSTOMER_INFO.chat_id,
-          msg_type: "text",
-          reply_in_thread: true,
-          content: JSON.stringify({
-            text: content
-          })
-        }
-      });
-
-      if ((content && replyResponse.msg === "success") || (isSendImagesSuccess.every(Boolean))) {
-        // Update
-        await retryQuery(() => this.db.$transaction(async (tx) => {
-          return tx.erpnextSalesOrderNotificationTracking.updateMany({
-            where: {
-              uuid: notificationTracking.uuid
-            },
-            data: {
-              order_data: {
-                items: salesOrderData.items,
-                attachments: salesOrderData.attachments,
-                paid_amount: salesOrderData.paid_amount,
-                deposit_amount: salesOrderData.deposit_amount
-              }
-            }
-          });
-        }, dbConnection));
-        return { success: true, message: "Gửi cập nhật đơn thành công!" };
-      }
-
-      return { success: false, message: "Đơn hàng này đã được gửi thông báo từ trước đó!" };
-    }
-
-    if (isUpdateMessage) {
-      return { success: true, message: "Ok" };
-    }
-
-    const content = await this.composeNewOrderContent(salesOrderData, customer, promotionData);
-
-    const _response = await larkClient.im.message.create({
-      params: {
-        receive_id_type: "chat_id"
-      },
-      data: {
-        receive_id: CHAT_GROUPS.CUSTOMER_INFO.chat_id,
-        msg_type: "text",
-        content: JSON.stringify({
-          text: content
-        })
-      }
-    });
-
-    const messageId = _response.data.message_id;
-
-    if (salesOrderData.attachments && salesOrderData.attachments.length > 0) {
-      await this._sendAttachmentsToLark(
-        larkClient,
-        salesOrderData.attachments,
-        messageId,
-        CHAT_GROUPS.CUSTOMER_INFO.chat_id
-      );
-    }
-
-    await retryQuery(() => this.db.erpnextSalesOrderNotificationTracking.create({
-      data: {
-        lark_message_id: messageId,
-        order_name: salesOrderData.name,
-        haravan_order_id: salesOrderData.haravan_order_id,
-        order_data: {
-          items: salesOrderData.items,
-          attachments: salesOrderData.attachments,
-          paid_amount: salesOrderData.paid_amount,
-          deposit_amount: salesOrderData.deposit_amount
-        }
-      }
-    }));
-
-    return { success: true, message: "Đã gửi thông báo thành công!" };
+  async sendNotificationToLark(orderData, isUpdateMessage = false) {
+    return sendSalesOrderNotificationToLark(this.frappeClient, this.db, this.env, orderData, isUpdateMessage);
   }
 
   async syncSalesOrdersToDatabase(options = {}) {
@@ -622,235 +315,7 @@ export default class SalesOrderService {
     }
   }
 
-  async composeUpdateOrderContent(oldSalesOrderData, salesOrderData, promotionData) {
-    return composeOrderUpdateMessage(oldSalesOrderData, salesOrderData, promotionData);
-  }
-
-  async _getLarkUserIdByEmail(email) {
-    if (!email) return null;
-    const user = await this.db.larksuite_users.findFirst({
-      where: {
-        OR: [
-          { email: email },
-          { enterprise_email: email }
-        ]
-      },
-      select: { user_id: true }
-    });
-    return user?.user_id || null;
-  }
-
-  async composeNewOrderContent(salesOrderData, orderCustomer, promotionData, isReorder = false) {
-    const customer = orderCustomer ?? (await this.frappeClient.getDoc("Customer", salesOrderData.customer));
-
-    const leadSource = await this.frappeClient.getDoc("Lead Source", customer.first_source);
-
-    const policyNames = salesOrderData.policies.map(policy => policy.policy);
-    const policyData = await this.frappeClient.getList("Policy", {
-      filters: [["name", "in", policyNames]]
-    });
-
-    const productCategoryNames = salesOrderData.product_categories.map(productCategory => productCategory.product_category);
-    const productCategoryData = await this.frappeClient.getList("Product Category", {
-      filters: [["name", "in", productCategoryNames]]
-    });
-
-    const purposeNames = (salesOrderData.sales_order_purposes || []).map(purpose => purpose.purchase_purpose || purpose.sales_order_purpose || purpose.purpose);
-    const purposeData = await this.frappeClient.getList("Purchase Purpose", {
-      filters: [["name", "in", purposeNames]]
-    });
-
-    const primarySalesPersonName = salesOrderData.primary_sales_person;
-    const primarySalesPerson = await this.frappeClient.getDoc("Sales Person", primarySalesPersonName);
-
-    const secondarySalesPersonNames = salesOrderData.sales_team
-      .filter(salesPerson => salesPerson.sales_person !== salesOrderData.primary_sales_person)
-      .map(salesPerson => salesPerson.sales_person);
-
-    const secondarySalesPeople = await this.frappeClient.getList("Sales Person", {
-      filters: [["name", "in", secondarySalesPersonNames]]
-    });
-
-    let content = composeSalesOrderNotification(salesOrderData, promotionData, leadSource, policyData, productCategoryData, purposeData, customer, primarySalesPerson, secondarySalesPeople);
-
-    if (isReorder) {
-      const hasMissingSerial = salesOrderData.items?.some(item => isMissingJewelrySerial(item));
-      if (hasMissingSerial) {
-        const primarySalesUserId = await this._getLarkUserIdByEmail(primarySalesPerson?.employee_email);
-        if (primarySalesUserId) {
-          content += `\n* <b>Thiếu thông tin:</b>\n- <at user_id="${primarySalesUserId}"></at> Vui lòng bổ sung số serial cho các sản phẩm trang sức còn thiếu!\n\n`;
-        }
-      }
-    }
-
-    return content;
-  }
-
-  async _sendAttachmentsToLark(larkClient, attachments, messageId, chatId) {
-    if (!attachments || attachments.length === 0) {
-      return [];
-    }
-
-    return await Promise.all(
-      attachments.map(async (file) => {
-        if (file.is_private) {
-          const replyResponse = await larkClient.im.message.reply({
-            path: {
-              message_id: messageId
-            },
-            data: {
-              receive_id: chatId,
-              msg_type: "text",
-              reply_in_thread: true,
-              content: JSON.stringify({
-                text: file.file_url
-              })
-            }
-          });
-          return replyResponse.msg === "success";
-        } else {
-          const r2Key = SalesOrderService._extractR2KeyFromUrl(file.file_url);
-          if (!r2Key) return false;
-          const imageBuffer = await new ERPR2StorageService(this.env).getObjectByKey(r2Key);
-          if (!imageBuffer) return false;
-          return Larksuite.Messaging.ImageMessagingService.sendLarkImageFromUrl({
-            imageBuffer,
-            chatId: chatId,
-            env: this.env,
-            rootMessageId: messageId
-          });
-        }
-      })
-    );
-  }
-
-  static _extractR2KeyFromUrl(url) {
-    try {
-      const urlObj = new URL(url);
-      const r2KeyParam = urlObj.searchParams.get("key");
-      if (r2KeyParam) {
-        return decodeURIComponent(r2KeyParam);
-      }
-      if (urlObj.pathname.length > 1) {
-        return decodeURIComponent(urlObj.pathname.substring(1));
-      }
-      return null;
-    } catch (e) {
-      Sentry.captureException(e);
-      return null;
-    }
-  }
-
   async syncHaravanFinancialStatus(salesOrderData) {
-    if (!salesOrderData.haravan_order_id) {
-      return;
-    }
-
-    if (Math.abs(salesOrderData.grand_total - salesOrderData.paid_amount) <= 1000) {
-      const HRV_API_KEY = this.env.HARAVAN_TOKEN;
-      if (!HRV_API_KEY) {
-        return;
-      }
-      const haravanClient = new HaravanAPI(HRV_API_KEY);
-      try {
-        const response = await haravanClient.order.getOrder(salesOrderData.haravan_order_id);
-        const haravanOrder = response.order;
-
-        if (haravanOrder.financial_status === "paid") {
-          return;
-        }
-
-        const { remainingBalance: remainingAmount } = getOrderFinancials(haravanOrder);
-
-        if (remainingAmount > 0) {
-          await haravanClient.orderTransaction.createTransaction(salesOrderData.haravan_order_id, {
-            amount: remainingAmount,
-            kind: "capture",
-            gateway: SalesOrderService.PAYMENT_GATEWAY_ERP
-          });
-        }
-      } catch (error) {
-        Sentry.captureException(error);
-      }
-    }
+    return syncHaravanFinancialStatus(this.env, salesOrderData);
   }
-
-  async getAllRelatedSalesOrders(initialOrderName, initialOrderDoc = null) {
-    const relatedOrdersMap = new Map();
-    const initialOrder = initialOrderDoc || await this.frappeClient.getDoc("Sales Order", initialOrderName);
-
-    if (!initialOrder) return {
-      allRelatedOrders: []
-    };
-
-    relatedOrdersMap.set(initialOrderName, {
-      name: initialOrder.name,
-      cancelled_status: initialOrder.cancelled_status,
-      split_order_group: initialOrder.split_order_group
-    });
-
-    if (initialOrder.is_split_order && initialOrder.split_order_group) {
-      const groupOrders = await this.frappeClient.getList("Sales Order", {
-        filters: [
-          ["split_order_group", "=", initialOrder.split_order_group],
-          ["is_split_order", "=", 1]
-        ],
-        fields: ["name", "cancelled_status", "split_order_group"]
-      });
-
-      groupOrders.forEach(o => relatedOrdersMap.set(o.name, {
-        name: o.name,
-        cancelled_status: o.cancelled_status,
-        split_order_group: o.split_order_group
-      }));
-    }
-
-    const toVisit = Array.from(relatedOrdersMap.keys());
-    const visited = new Set(relatedOrdersMap.keys());
-
-    while (toVisit.length > 0) {
-      const currentSoName = toVisit.pop();
-      let currentOrderDoc = null;
-      if (currentSoName === initialOrderName) {
-        currentOrderDoc = initialOrder;
-      } else {
-        try {
-          currentOrderDoc = await this.frappeClient.getDoc("Sales Order", currentSoName);
-        } catch (e) {
-          console.warn(`Could not fetch Sales Order ${currentSoName} for traversal`, e);
-          continue;
-        }
-      }
-
-      if (currentOrderDoc) {
-        relatedOrdersMap.set(currentSoName, currentOrderDoc);
-      }
-
-      if (currentOrderDoc && currentOrderDoc.ref_sales_orders) {
-        for (const ref of currentOrderDoc.ref_sales_orders) {
-          if (ref.sales_order && !visited.has(ref.sales_order)) {
-            visited.add(ref.sales_order);
-            toVisit.push(ref.sales_order);
-          }
-        }
-      }
-
-      const refsUp = await this.frappeClient.getList("Sales Order", {
-        filters: [["Sales Order Reference", "sales_order", "=", currentSoName]],
-        fields: ["name"]
-      });
-
-      for (const parentOrder of refsUp) {
-        if (parentOrder.name && !visited.has(parentOrder.name)) {
-          visited.add(parentOrder.name);
-          toVisit.push(parentOrder.name);
-        }
-      }
-    }
-
-    return {
-      allRelatedOrders: Array.from(relatedOrdersMap.values())
-    };
-  }
-
 }
